# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is an Car sales api.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "AutoMart API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "ezeugwukenneth42@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "herokuapp.kennymorgan1.com",
  "basePath": "/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/car": {
      "post": {
        "tags": [
          "car"
        ],
        "summary": "Add a new car AD",
        "description": "",
        "operationId": "postCar",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "car object that needs to be advertised",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Car"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/CarSuccess"
            }
          },
          "400": {
            "description": "Invalid car detail supplied"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
...
    "/order": {
      "post": {
        "tags": [
          "order"
        ],
        "summary": "place order",
        "description": "",
        "operationId": "purchaseOrder",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "car object that needs to be advertised",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/OrderSuccess"
            }
          },
          "400": {
            "description": "Invalid car detail supplied"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Order": {
      "type": "object",
      "properties": {
        "car_id": {
          "type": "integer",
          "example": 1
        },
        "price": {
          "type": "integer",
          "example": 10000
        },
        "status": {
          "type": "array",
          "xml": {
            "name": "tag",
            "wrapped": true
          },
          "items": {
            "type": "string"
          },
          "example": [
            "pending",
            "accepted",
            "rejected"
          ]
        }
      },
      "xml": {
        "name": "Order"
      }
    },
    "OrderSuccess": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "car_id": {
          "type": "integer",
          "example": 1
        },
        "created_on": {
          "type": "string",
          "example": "12-12-2018"
        },
        "status": {
          "type": "string",
          "example": "pending"
        },
        "price": {
          "type": "integer",
          "example": 10000
        },
        "price_offered": {
          "type": "integer",
          "example": 10000
        }
      }
    },
...
    "Car": {
      "type": "object",
      "properties": {
        "state": {
          "type": "array",
          "xml": {
            "name": "tag",
            "wrapped": true
          },
          "items": {
            "type": "string"
          },
          "example": [
            "used",
            "new"
          ]
        },
        "price": {
          "type": "integer",
          "example": 5000000
        },
        "manufacturer": {
          "type": "string",
          "example": "Toyota"
        },
        "model": {
          "type": "string",
          "example": 2018
        },
        "body_type": {
          "type": "string",
          "example": "car"
        }
      }
    },
    "CarSuccess": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "email": {
          "type": "string",
          "example": "ezeugwukenneth@automart.com"
        },
        "created_on": {
          "type": "string",
          "example": "12-12-2018"
        },
        "manufacturer": {
          "type": "string",
          "example": "Toyota"
        },
        "model": {
          "type": "string",
          "example": 2018
        }
      }
    },
...
  },
...
}